company_registry_orig/requirements.txt
Django==3.0.8
djangorestframework==3.11.0
fuzzywuzzy==0.18.0
